Sprite animation is now done with the SpriteAnimatorSystem
