--- Source/MonoGame.Extended/Entities/Systems/AnimatedSpriteSystem.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.Xna.Framework;
-using MonoGame.Extended.Sprites;
-
-namespace MonoGame.Extended.Entities.Systems
-{
-    public class AnimatedSpriteSystem : EntitySystem
-    {
-        protected override void Update(Entity entity, GameTime gameTime)
-        {
-            var deltaTime = gameTime.GetElapsedSeconds();
-            foreach (AnimatedSprite sprite in entity.GetComponents<Sprite>())
-                sprite.Update(deltaTime);
-        }
-    }
-}
